requirements.txt
asgiref==3.7.2
colorama==0.4.6
colorlog==6.7.0
APScheduler==3.10.4
et-xmlfile==1.1.0
openpyxl==3.1.2
pyodbc==5.0.1
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
SQLAlchemy~=2.0.27
pandas~=2.2.2
alembic~=1.13.1
pydantic~=2.6.3
fastapi==0.110.2
celery==5.4.0
gunicorn==22.0.0
flower==2.0.1
uvicorn==0.29.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-json-logger==2.0.7
psutil==5.9.7






rapidfuzz==3.9.3
Django==4.2.6
django-filter==23.3